Crate domain_pallet_executive

Domain Executive Module

This module is derived from frame_executive with some custom modifications for collecting the intermediate storage roots in the block execution required for the fraud proof of decoupled execution in Subspace.

Structs

Executive

Same semantics with frame_executive::Executive.

Pallet

The pallet implementing the on-chain logic.

Enums

Call

Contains one variant per dispatchable that can be called by an extrinsic.

Event

The event emitted by this pallet.

Traits

Config

Configuration trait of this pallet.

Type Definitions

CallOf

CheckedOf

ModuleDeprecated

Type alias to Pallet, to be used by construct_runtime.

OriginOf