domain_pallet_executive
Trait Config

pub trait Config: Config {
    type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>;
    type WeightInfo: WeightInfo;
    type Currency: Mutate<Self::AccountId>;
    type LengthToFee: WeightToFee<Balance = <<Self as Config>::Currency as Inspect<<Self as Config>::AccountId>>::Balance>;
    type ExtrinsicStorageFees: ExtrinsicStorageFees<Self>;
}

Configuration trait of this pallet.

The main purpose of this trait is to act as an interface between this pallet and the runtime in which it is embedded in. A type, function, or constant in this trait is essentially left to be configured by the runtime that includes this pallet.

Consequently, a runtime that wants to include this pallet must implement this trait.

Required Associated Types

type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>

type WeightInfo: WeightInfo

type Currency: Mutate<Self::AccountId>

type LengthToFee: WeightToFee<Balance = <<Self as Config>::Currency as Inspect<<Self as Config>::AccountId>>::Balance>

type ExtrinsicStorageFees: ExtrinsicStorageFees<Self>

Object Safety

This trait is not object safe.