domain_pallet_executive
Trait ExtrinsicStorageFees

pub trait ExtrinsicStorageFees<T: Config> {
    // Required methods
    fn extract_signer(
        xt: <<T as Config>::Block as BlockT>::Extrinsic,
    ) -> (Option<<T as Config>::AccountId>, DispatchInfo);
    fn on_storage_fees_charged(
        charged_fees: <<T as Config>::Currency as Inspect<<T as Config>::AccountId>>::Balance,
        tx_size: u32,
    ) -> Result<(), TransactionValidityError>;
}

Trait trait used to charge the extrinsic storage.

Required Methods

fn extract_signer( xt: <<T as Config>::Block as BlockT>::Extrinsic, ) -> (Option<<T as Config>::AccountId>, DispatchInfo)

Extracts signer from given extrinsic and its dispatch info.

fn on_storage_fees_charged( charged_fees: <<T as Config>::Currency as Inspect<<T as Config>::AccountId>>::Balance, tx_size: u32, ) -> Result<(), TransactionValidityError>

Hook to note operator rewards for charged storage fees.

Object Safety

This trait is not object safe.