Crate domain_pallet_executive

Domain Executive Module

This module is derived from frame_executive with some custom modifications for collecting the intermediate storage roots in the block execution required for the fraud proof of decoupled execution in Subspace.

Re-exports

pub use __tt_error_token_1 as tt_error_token;

pub use __tt_default_parts_7 as tt_default_parts;

pub use __tt_extra_parts_7 as tt_extra_parts;

pub use __tt_default_parts_v2_7 as tt_default_parts_v2;

Modules

dispatchables

Auto-generated docs-only module listing all defined dispatchables for this pallet.

storage_types

Auto-generated docs-only module listing all (public and private) defined storage types for this pallet.

weights

Autogenerated weights for domain_pallet_executive

Structs

Executive

Same semantics with frame_executive::Executive.

Pallet

The Pallet struct, the main type that implements traits and standalone functions within the pallet.

Enums

Call

Contains a variant per dispatchable extrinsic that this pallet has.

Event

The Event enum of this pallet

Traits

Config

Configuration trait of this pallet.

ExtrinsicStorageFees

Trait trait used to charge the extrinsic storage.

Type Aliases

CallOf

CheckedOf

ModuleDeprecated

Type alias to Pallet, to be used by construct_runtime.

OriginOf